+++ Components/UnitFrames/UnitFrames.lua
@@ -1,29 +1,4 @@
---[[
-
-	The MIT License (MIT)
-
-	Copyright (c) 2022 Lars Norberg
-
-	Permission is hereby granted, free of charge, to any person obtaining a copy
-	of this software and associated documentation files (the "Software"), to deal
-	in the Software without restriction, including without limitation the rights
-	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-	copies of the Software, and to permit persons to whom the Software is
-	furnished to do so, subject to the following conditions:
-
-	The above copyright notice and this permission notice shall be included in all
-	copies or substantial portions of the Software.
-
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-	SOFTWARE.
-
---]]
-local Addon, ns = ...
+﻿local Addon, ns = ...
 local UnitFrames = ns:NewModule("UnitFrames", "LibMoreEvents-1.0")
 local oUF = ns.oUF
 
@@ -631,11 +606,13 @@ UnitFrames.OnInitialize = function(self)
 
 	self:SpawnUnitFrames()
 	self:SpawnGroupFrames()
-	-- Spawn nameplates only if enabled in settings
+	-- Custom nameplates disabled - using Blizzard default or other addon
+	--[[ Spawn nameplates only if enabled in settings
 	local db = ns.db
 	if db and db.global and db.global.unitframes and db.global.unitframes.enableNamePlates then
 		self:SpawnNamePlates()
 	end
+	--]]
 end
 
 UnitFrames.OnEnable = function(self)